feat: extract kwil infra
