+++ packages/@core/src/index.ts
@@ -4,6 +4,7 @@ export * from "./cryptography";
 export * from "./types";
 export * from "./kwil-nep413-signer";
 export * from "./kwil-actions";
+export * from "./kwil-infra";
 export * from "./store";
 export * from "./utils";
 export * from "./enclave";
