+++ packages/@core/src/kwil-actions/credentials.ts
@@ -1,5 +1,5 @@
+import type { KwilActionClient } from "../kwil-infra";
 import type { InsertableIDOSCredential, idOSCredential, idOSGrant } from "../types";
-import type { KwilActionClient } from "./create-kwil-client";
 
 export interface CreateCredentialParams {
   user_id: string;
