+++ packages/@core/src/kwil-actions/index.ts
@@ -1,6 +1,4 @@
 export * from "./attributes";
-export * from "./create-kwil-client";
-export * from "./create-kwil-signer";
 export * from "./credentials";
 export * from "./grants";
 export * from "./user";
