+++ packages/@core/src/kwil-infra/create-kwil-client.ts
@@ -2,7 +2,7 @@ import { type KwilSigner, NodeKwil, WebKwil } from "@kwilteam/kwil-js";
 import type { Config } from "@kwilteam/kwil-js/dist/api_client/config";
 import type { ActionBody, CallBody, PositionalParams } from "@kwilteam/kwil-js/dist/core/action";
 import invariant from "tiny-invariant";
-import { actionSchema } from "./schema";
+import { actionSchema } from "../kwil-actions/schema";
 
 interface CreateKwilClientParams {
   chainId?: string;
