+++ packages/@core/src/kwil-infra/create-near-wallet-kwil-signer.ts
@@ -11,9 +11,9 @@ import {
   bytesConcat,
   utf8Decode,
 } from "../codecs";
+import type { KwilActionClient } from "../kwil-infra/create-kwil-client";
 import { implicitAddressFromPublicKey } from "../kwil-nep413-signer";
 import type { Store } from "../store";
-import type { KwilActionClient } from "./create-kwil-client";
 
 const NEAR_WALLET_TYPES: string[] = [
   "browser",
